Fix arg type declaration mismatch
- Now that we're trying to compile 64-bit these differences
  are significant.
- Fixes a warning.

+++ CustomControls.m
@@ -403,7 +403,7 @@
     _selectedTitleColor = [selectedTitleColor retain];
 }
 
-- (void)drawSegment:(int)segment inFrame:(NSRect)frame withView:(NSView*)controlView
+- (void)drawSegment:(NSInteger)segment inFrame:(NSRect)frame withView:(NSView*)controlView
 {
     NSColor* titleColor = ([self selectedSegment] == segment) ?
                                             _selectedTitleColor : _titleColor;
